[refactor] add trap.rs
